oh, it's done
Honestly, I was expecting it to be a lot harder than this lol, this was supposed to be my "rough test" version but it just kinda works... I set aside a whole day to work on this and it's done. I feel empty.

Just kidding I've still gotta work on the string stuff, I have more work to do, but I am still surprised at how easily it all just kinda worked without needing to rework the movement.